fetch contributors from github api
